filter-test-logs should signal failure when the scripts fail partway through https://bugs.webkit.org/show_bug.cgi?id=290911
Reviewed by NOBODY (OOPS!).

run-jsc-stress-tests or run-javascriptcore-tests may run into an
exception and fail to print the results. In that case, filter-test-logs
should raise the alarm.

Change filter-test-logs to both exit with a non-zero status and
pass-through the last 500 lines of the output when that happens (to
allow for easier debugging).

* Tools/Scripts/filter-test-logs:
(filter_jsc_tests):
(filter_layout_tests):
(filter_scan_build):
(filter_webdriver_tests):
(main):