[CISupport][GTK][WPE]: Allow passing a reference local mirror repository when cloning the WebKit git repository https://bugs.webkit.org/show_bug.cgi?id=291390
Reviewed by NOBODY (OOPS!).

We are starting to experiment deploying workers on containers without permanent disk storage
which means that the disk of the worker will be wiped on reboot/shutdown, so when it starts
it will have to clone the WebKit repository from scratch and this can take up to 10 minutes
even with a good Internet connection.

So to speed up this process we plan to keep a local git mirror mounted at some local path
and then clone the WebKit repository by passing to the command `git clone` the arguments:
`--reference /path/to/WebKit.git`.

Buildbot class `buildbot.steps.source.git.Git` supports this, is just a matter of defining
the class variable `self.reference` to this path and it will take care of it.

This patch allows any GTK/WPE/JSCOnly worker to enable this by simple defining on the
worker own environment the variable `WEBKIT_MIRROR_PATH=/path/to/WebKit.git`.
This allows flexibility to dinamically change the path for each worker at run-time
without needing to hardcode the value in the master config neither restarting the master.

This environment variable from the worker is evaluated at the step `PrintAndSetConfiguration()`
which always runs before the step `CheckOutSource()`. That way we ensure that the property
defining the optional path to this local mirror is defined when `git clone` runs.

* Tools/CISupport/build-webkit-org/steps.py:
(CheckOutSource.run):
(PrintConfiguration):
(PrintConfiguration.getResultSummary):
* Tools/CISupport/build-webkit-org/steps_unittest.py:
* Tools/CISupport/ews-build/steps.py:
(CheckOutSource.run):
(PrintConfiguration):
(PrintConfiguration.run):
(PrintConfiguration.parseAndValidateApple):
(PrintConfiguration.parseAndValidate): Deleted.
* Tools/CISupport/ews-build/steps_unittest.py: