Position Model UITargetedDragPreview for Drag and Drop https://bugs.webkit.org/show_bug.cgi?id=291468 rdar://146525262
Reviewed by NOBODY (OOPS!).

This PR introduces a new API to animate the model to fit within the portal bounds in preparation
for drag and drop before we create a UITargetedDragPreview for it. It uses the orbit-fit position
such that if the model's bounds exceeds the portal's bounds, it will then animate to the orbit fit
location before the lift animation occurs.

* Source/WebKit/WebKitSwift/RealityKit/RKEntity.swift:
(WKSRKEntity.preFitAnimationTransformMatrix):
(WKSRKEntity.preFitAnimationController):
(WKSRKEntity.preFitAnimationCompletionSubscription):
(WKSRKEntity.ignoreTransformUpdates):
(WKSRKEntity.fitEntity(within:animated:completion:)):
(WKSRKEntity.resetModelTransformAfterDrag):
* Source/WebKit/WebKitSwift/RealityKit/RealityKitBridging.h: