Re-introduce container entity for standardizing model transform pivot https://bugs.webkit.org/show_bug.cgi?id=291605 rdar://149343680
Reviewed by NOBODY (OOPS!).

This PR reintroduces the container entity to the model element view hierarchy. This
was previously removed as part of a refactor to move parts of the model to the UI Process.
However, in an effort to standardize the pivot point for positioning, rotating, and scaling
the model, we are re-adding it back into the hierarchy such that it is positioned at
the initial bounding box center of the model. We make the container an RKEntity so that
we can apply transforms and reparent it while preserving world transforms. We also clean up
some unused APIs as a result of this change.

* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.h:
* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.mm:
(WebKit::ModelProcessModelPlayerProxy::updateTransform):
(WebKit::ModelProcessModelPlayerProxy::didFinishLoading):
(WebKit::ModelProcessModelPlayerProxy::setStageMode):
(WebKit::ModelProcessModelPlayerProxy::updateTransformSRT):
* Source/WebKit/WebKitSwift/RealityKit/RKEntity.swift:
(WKSRKEntity.interactionPivotPoint):
(WKSRKEntity.interactionContainerDidRecenter(_:)):
(WKSRKEntity.recenterEntity(at:)): Deleted.
* Source/WebKit/WebKitSwift/RealityKit/RealityKitBridging.h:
* Source/WebKit/WebKitSwift/StageMode/WKStageMode.h:
* Source/WebKit/WebKitSwift/StageMode/WKStageMode.swift:
(WKStageModeInteractionDriver.interactionContainerRef): Deleted.