ASAN_ILL | WebCore::CSSAnimation::syncStyleOriginatedTimeline https://bugs.webkit.org/show_bug.cgi?id=291601 rdar://147939911
Reviewed by NOBODY (OOPS!).

The issue was caused by change in animation mode from CSS to Web based in the
included http test script. The web based annimation does not make use of
owning element and delets the same. This causes failure when syncStyleOriginatedTimeline
function is called. The fix is to have a check for owningElement before calling the above
function.

The syncStyleOriginatedTimeline had a unused variable target and that was also
cleaned up

* LayoutTests/animations/cssAnimToWebAnimSwitchCrash-expected.txt: Added.
* LayoutTests/animations/cssAnimToWebAnimSwitchCrash.html: Added.
* Source/WebCore/animation/CSSAnimation.cpp:
(WebCore::CSSAnimation::syncPropertiesWithBackingAnimation):
(WebCore::CSSAnimation::syncStyleOriginatedTimeline):