In <model> 1 spatial centimeter should be equal to 1 css centimeter https://bugs.webkit.org/show_bug.cgi?id=291654 rdar://148281742
Reviewed by NOBODY (OOPS!).

This PR adds two static functions to convert the entityTransform coordinate space such that
the scale and position are standardized in a way where 1 spatial cm = 1 css cm. This means
that a 1cm cube should have a scale factor of 1.0 when it is placed in a portal with width
and height of 1 css cm.

* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.mm:
(WebKit::modelStandardizedTransformSRT):
(WebKit::modelLocalizedTransformSRT):
(WebKit::ModelProcessModelPlayerProxy::computeTransform):
(WebKit::ModelProcessModelPlayerProxy::setEntityTransform):
(WebKit::ModelProcessModelPlayerProxy::updateTransformSRT):