Remove Portal Crossing component when drag ends https://bugs.webkit.org/show_bug.cgi?id=291783 rdar://149234418
Reviewed by NOBODY (OOPS!).

Previously, model elements should get clipped by the front of the portal plane when the model extends
past the portal plane. With the portal crossing component, the clipping behavior has changed. To
restore the original behavior, we need to remove the component entirely when we disable the crossing
component.

* Source/WebKit/UIProcess/ios/WKPageHostedModelView.mm:
(-[WKPageHostedModelView init]):
(-[WKPageHostedModelView setPortalCrossing:]):