Wrong 'auto' track size in minmax content for a grid item with a specified 'width:max-content' https://bugs.webkit.org/show_bug.cgi?id=216144
Reviewed by NOBODY (OOPS!).

The addition of the new condition in sizeTrackToFitNonSpanningItem allows the grid track to grow to the max-content contribution of its items if the track has an auto min track sizing function and the grid container is being sized under a min/max-content constraint. This behavior applies only when the min track sizing function is auto.

* LayoutTests/fast/css-grid-layout/max-content-track-sizing-expected.txt: Added.
* LayoutTests/fast/css-grid-layout/max-content-track-sizing.html: Added.
* Source/WebCore/rendering/GridTrackSizingAlgorithm.cpp:
(WebCore::GridTrackSizingAlgorithm::isUnderMaxContentConstraint const):
(WebCore::GridTrackSizingAlgorithm::sizeTrackToFitNonSpanningItem):
* Source/WebCore/rendering/GridTrackSizingAlgorithm.h: