1 cm cube doesn't fit fully into a 1 css cm portal https://bugs.webkit.org/show_bug.cgi?id=292335 rdar://149429492
Reviewed by NOBODY (OOPS!).

Currently, when applying object fit for a model into a portal, there is still a gap between the model
and the portal bounds. This is happening because the bounds we return to compute the model's fit does
not factor in the `minimumEffectiveDeviceWidthScale` which scales up the viewport's 'device width' and
thus affects the total meters that the point size occupies. Therefore, we also need to account for that
scale when computing the model's 'fit scale'.

* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.mm:
(WebKit::makeMeterSizeFromPointSize):