WebCore::RenderObject::repaintUsingContainer ASSERT due to layer being nullptr https://bugs.webkit.org/show_bug.cgi?id=292353 rdar://148249823
Reviewed by NOBODY (OOPS!).

The ASSERT was triggered by repaintContainer->isComposited() being false. This
function returns false because hasLayer() returns false, which is correct since
the layer has been set to nullptr when the object in that layer was removed. The
fix is to move the ASSERT after the if condition to check if the layer is valid
and not nullptr.

* LayoutTests/compositing/compositing-deleted-object-assert-expected.txt: Added.
* LayoutTests/compositing/compositing-deleted-object-assert.html: Added.
* Source/WebCore/rendering/RenderObject.cpp:
(WebCore::RenderObject::repaintUsingContainer const):