[SOUP] WebsiteDataStore.ignoreTLSErrors is initialized to true when created via C API https://bugs.webkit.org/show_bug.cgi?id=292752
Reviewed by NOBODY (OOPS!).

Changed default value of m_ignoreTLSErrors to false, so that the
errors are not ignored by default if the data store has been created
via the C API.

* Source/WebKit/UIProcess/API/C/WKWebsiteDataStoreRef.cpp:
(WKWebsiteDataStoreSetIgnoreTLSErrorsForTesting):
There is no dedicated C API for changing the property, so adding
one for testing.
* Source/WebKit/UIProcess/API/C/WKWebsiteDataStoreRef.h:
* Source/WebKit/UIProcess/WebsiteData/WebsiteDataStore.h:

* Tools/WebKitTestRunner/TestController.cpp:
Make sure that the TLS errors are still ignored in the tests.