Starting opportunistically deleting GlyphDisplayListCache entries. Need the bug URL (OOPS!). Include a Radar link (OOPS!).
Reviewed by NOBODY (OOPS!).

Removing entries from the GlyphDisplayListCache and destroying them can
be expensive in that it appears in traces for benchmarks when destroying
line layout. To avoid doing this work on the critical path, we can use
the OpportunisticTaskScheduler to destroy these objects at a later point
during idle time.

To accomplish this, patch, whenever we remove an entry from the cache, we
will append it to a list that will be opportunistically cleaned up. In
order to decrease the likelihood of increasing the memory footprint, we
place a maximum limit on the number of items that can be included in
this list. After looking at some benchmarks locally, I decided to use 300
items as this maximum. OTS will then come and do a bit of cleanup by
removing min(entries.size(), 100) before moving onto other tasks. This
is to avoid hogging all of the time and allow other code aware of the
OTS deadline to execute.