[JSC] Make wasm-omg stress mode always tier up to OMG https://bugs.webkit.org/show_bug.cgi?id=293437 rdar://151866584
Reviewed by NOBODY (OOPS!).

The old config enabled the interpreters which was preventing
tier up to OMG on the very first execution. Disabling
IPInt and LLInt means we tier up through BBQ and into OMG
immediately, given the other test options.

Testing: added an assert(false) to OMGIRGenerator::addRefI31 and
ran a test that does a single ref.i31. Verified that without
this change, the assert is not triggered. With this change,
the assert triggers.

* Tools/Scripts/run-jsc-stress-tests: