[GStreamer][MSE] Support playing videos of less than 1 sec with libav decoders https://bugs.webkit.org/show_bug.cgi?id=293239
Reviewed by NOBODY (OOPS!).

GStreamer video decoders based on libav are currently implemented in a
way that hoard one video frame per decoding thread. This not only means
more latency during continuous playback, but also on stalls (when no
more frames are supplied at the video decoder input). While problem has
been mitigated by reducing the max-threads to 2 [1], the behaviour still
causes problems in some tests with very short videos.

This commit checks that we're currently using the libav video decoder
and, if the video position hasn't moved but in theory should, emits a
still frame event to force the video decoder to emit all the video
frames at its source pad.

A layout test to check that the video decoder has emitted all the frames
when a stall happens has been added. This test explicitly doesn't check
the exact played timestamps, to not fail because of missing edit list
support, which skews the timestamps.

See also:
https://bugs.webkit.org/show_bug.cgi?id=270622
https://bugs.webkit.org/show_bug.cgi?id=231019

[1] https://github.com/WebKit/WebKit/commit/128e962baa61f49c78087a622eaecd08dc692ef6#diff-4376b854280205494dc9de0f08f985afd4d6584a7487c0a136adac2954c003cdR2831

* LayoutTests/media/media-source/media-source-shortmediastall-expected.txt: Added.
* LayoutTests/media/media-source/media-source-shortmediastall.html: Added.
* Source/WebCore/platform/graphics/gstreamer/GStreamerCommon.h: Added the LibAv video decoding platform.
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::playbackPosition const): Emit the still frame event if the video hasn't advanced but in theory should have, but only when using a libav decoder and when the player isn't stalled in the same timestamp as the last time (avoid redundant emission).
(WebCore::MediaPlayerPrivateGStreamer::configureVideoDecoder): Declare that we're using the LibAv video decoding platform when present.
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.h: Added m_lastStalledPosition to store the timestamp where the player got last stalled. This is used to avoid redundant emission of the still frame event on consecutive stalls that are in fact for the the same timestamp.
* Source/WebCore/platform/graphics/gstreamer/mse/MediaPlayerPrivateGStreamerMSE.cpp:
(WebCore::MediaPlayerPrivateGStreamerMSE::hasFutureTime const): Expose MediaSourcePrivate::hasFutureTime() through the player private.
* Source/WebCore/platform/graphics/gstreamer/mse/MediaPlayerPrivateGStreamerMSE.h: Add hasFutureTime().
* Source/WebCore/platform/graphics/gstreamer/mse/WebKitMediaSourceGStreamer.cpp:
(webKitMediaSrcSendEvent): Forward the GstEventStillFrame custom downstream event to the first (and only, if present) video src pad, but enqueing it in the buffer/event queue in MediaSourceTrackGStreamer.