[Web Inspector] Fix Need to press Tab twice to apply suggestion in Console https://bugs.webkit.org/show_bug.cgi?id=293704
Reviewed by NOBODY (OOPS!).

Autocompletion sometimes required pressing the Tab key twice to work.
This happened because, when matching the prefix with possible completions,
Rhe code did not normalize the prefix and completion values to either uppercase or lowercase.
So, while typing doc, for example, the Doc prefix from Document was being treated as a different prefix.
To fix this bug, this patch apply lowercase to prefix and completions before comparing them.

* Source/WebInspectorUI/UserInterface/Controllers/CodeMirrorCompletionController.js:
(WI.CodeMirrorCompl etionController.prototype._generateJavaScriptCompletions):