[JSC] Stop regenerating conditions for loops https://bugs.webkit.org/show_bug.cgi?id=294171 rdar://problem/152785322
Reviewed by NOBODY (OOPS!).

We are spreading a loop in bytecode in this way right now,

    init
    condition-check
    ...
    loop-body
    ...
    condition-check -> jump to loop-body

However, this causes a problem when loop-hint is placed in loop-body's
head. Since we will insert OSR entry, loop-body will see two
predecessors, in which one of that is not dominated by condition-check.
As a result, we fail to perform some optimizations (including integer
range optimizations) since body looks like non-dominated by
condition-check because of OSR entry block. So FTL failed to use
condition information to optimize the loop-body.

Anyway, generating condition twice in bytecode bloats bytecode so much
if condition is complicated. And if we would like to do the same thing,
we can do it in B3: duping condition block to the loop-body. So we
should stop doing it in bytecode level.

* Source/JavaScriptCore/bytecompiler/NodesCodegen.cpp:
(JSC::WhileNode::emitBytecode):
(JSC::ForNode::emitBytecode):