RenderGrid should use OrderIterator shouldSkipChild instead of querying renderer type directly https://bugs.webkit.org/show_bug.cgi?id=294637 rdar://problem/153679479
Reviewed by NOBODY (OOPS!).

In 246990@main, we added some logic to deal with a crash when a
RenderGrid was also a fieldset by skipping over the fieldset's legend at
different parts in grid layout as they do not seem to be part of the
formatting context's in flow layout.

Other renderers (e.g. RenderFlexibleBox) have this same type of logic
but instead use the OrderIterator's shouldSkipChild API. For at least
the sake of consistency, we should probably try to use the same function
instead of checking for isLegend() directly. Unfortunately, we cannot
simply just adopt them at the exact spot the isLegend() checks occur
during execution because the isExcludedFromNormalLayout bit only gets
set on renderers during the layoutExcludedChildren phase. So we first
need to move the computeAspectRatioDependentAndBaselineItems call to
under layoutExcludedChildren and then start using the function.