Replace UChar with char16_t https://bugs.webkit.org/show_bug.cgi?id=294673 rdar://153736656
Reviewed by NOBODY (OOPS!).

This replaces UChar with char16_t throughout the WebKit codebase. This was a
plain and simple textual substitution of \bUChar\b with char16_t.
The following directories were excluded:
* Source/WTF/icu
* Source/ThirdParty
* WebKitLibraries/SDKDBs/iphoneos18.2/libicucore.A.partial.sdkdb
* JSTests
* Source/WebKitLegacy/ios/Misc/WebUIKitSupport.h

The full file list has been excluded from this PR description because
otherwise it's too long and can't be uploaded. There are widespread changes in
the following directories:
* Source/JavaScriptCore
* Source/WTF
* Source/WebCore
* Source/WebGPU
* Source/WebKit
* Source/WebKitLegacy
* Tools/TestRunnerShared
* Tools/TestWebKitAPI
* Tools/gdb/webkit.py
* Tools/lldb/lldbWebKitTester