WebDriver BiDi: support for locating DOM nodes via `browsingContext.locateNodes` rdar://137019465
Reviewed by NOBODY (OOPS!).

Adding the majority of work for the locateNodes endpoint. The main focus is a need for a new JS
script as the classic implementation found as an atom couldn't be modified for the BIDI
specification. A current bug with the dispatcher https://bugs.webkit.org/show_bug.cgi?id=294633
has blocked end-to-end testing. The new script was individually tested against a test site,
ensuring each type of locator worked. Currently, the callback from evaluateJavaScriptFunction
is bare as it's difficult to create a thorough solution without testing.

* Source/WebKit/UIProcess/Automation/BidiBrowsingContextAgent.cpp:
(WebKit::BidiBrowsingContextAgent::locateNodes):
* Source/WebKit/UIProcess/Automation/BidiBrowsingContextAgent.h:
* Source/WebKit/UIProcess/Automation/protocol/BidiBrowsingContext.json:
* Source/WebKit/UIProcess/Automation/protocol/BidiScript.json: