Issue caused by vertical-align on root node with rem/rex type https://bugs.webkit.org/show_bug.cgi?id=294451 rdar://151618114
Reviewed by NOBODY (OOPS!).

The issue is when doing vertical-align of type rem or rex, it is assumed that there
is a valid root style. This is not so in case of root node and this was causing
issues (root style is NULL ofr root element).
The fix is in case the rootStyle is NULL, mark the property as invalid

* LayoutTests/fast/css/vertical-align-rem-on-root-expected.txt: Added.
* LayoutTests/fast/css/vertical-align-rem-on-root.html: Added.
* LayoutTests/platform/mac-wk1/TestExpectations:
* Source/WebCore/css/CSSPrimitiveValueMappings.h:
(WebCore::CSSPrimitiveValue::convertToLength const):