Disable CGStyle-based filters on google maps https://bugs.webkit.org/show_bug.cgi?id=295356 rdar://154317279
Reviewed by NOBODY (OOPS!).

CGStyle-based filters combined with large transforms cause
underpainting artifacts along some tile edges,
resulting in visible gaps between tiles.

On iOS devices you can see a flash of white line when transitioning
to a different location in Google Maps.
This issue does not occur on VMs or simulators.

We are temporarily disabling this feature while we
investigate the root cause of underpainting with CGStyle-based filters.