[EWS] Use a "pr_" prefix for the local branch name used to merge the changes from the PR https://bugs.webkit.org/show_bug.cgi?id=295420
Reviewed by NOBODY (OOPS!).

If someone sends a PR from a branch named "main" then the EWS will fail to checkout
correctly the PR because the EWS first tries to create a new branch locally with the
same name than the remote branch, but the name "main" is already reserved for the
main branch from origin.

This won't happen when the user uses the tool "git-webkit" because this tool renames
the user main branch to a branch named "eng/PATCH-TITLE", but it can happen if someone
sends a PR manually without using the tool "git-webkit"

This patch modifies the EWS to use a "pr_" prefix for the local branch name, in order
to avoid this conflict.

* Tools/CISupport/ews-build/steps.py:
(CheckOutPullRequest.getCommandList):
* Tools/CISupport/ews-build/steps_unittest.py: