[Win] WebKitTestRunnerWS not found running win-tests https://bugs.webkit.org/show_bug.cgi?id=296131
Reviewed by NOBODY (OOPS!).

We're running under mingw bash; need to convert to a windows style path
for the driver and imagediff paths. Move windows specific logic into the
windows port file.

Try explicitly setting the size of the window for WebKitTestRunnerWS,
this value will be used by CW_USEDEFAULT.

Try forcing the use of WARP when creating the display.