dNR: fix sub_frame resourceType allowAllRequests rules https://bugs.webkit.org/show_bug.cgi?id=295043 rdar://154124673
Reviewed by NOBODY (OOPS!).

This patch re-introduces the fixes that were made to fix sub_frame resourceType allowAllRequests
rules, which were reverted because they wouldn't work with site isolation. Now, instead of assuming
local frames when traversing the frame hierarchy, we use the frame's URL origin, which is also
available for remote frames. The downside to this is that sub_frame resourceType allowAllRequests
rules that want to filter based on other parts of the URL won't work. I've filed:

dNR: allowAllRequests with sub_frame resource type will only work with a URL filter acting on the
origin of the URL
https://bugs.webkit.org/show_bug.cgi?id=296294

Otherwise, this patch is exactly the same as before.

Also see:
https://bugs.webkit.org/show_bug.cgi?id=295541
https://bugs.webkit.org/show_bug.cgi?id=295036

* Source/WebCore/contentextensions/ContentExtensionCompiler.cpp:
* Source/WebCore/contentextensions/ContentExtensionParser.cpp:
* Source/WebCore/contentextensions/ContentExtensionRule.h:
* Source/WebCore/contentextensions/ContentExtensionsBackend.cpp:
* Source/WebCore/loader/ResourceLoadInfo.h:
* Source/WebKit/UIProcess/Extensions/Cocoa/_WKWebExtensionDeclarativeNetRequestRule.mm:
* Tools/TestWebKitAPI/Tests/WebCore/ContentExtensions.cpp:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtensionAPIDeclarativeNetRequest.mm: