Do not allow WorkerDedicatedRunLoop::run to spin tight loops that never handle tasks (message queue requests only ever timeout) rdar://155433911 https://bugs.webkit.org/show_bug.cgi?id=296163
Reviewed by NOBODY (OOPS!).

We have metrics showing ServiceWorkers getting caught in a tight loop of spinning over and over,
never actually handling a service worker task from the message queue, but instead firing a CFRunLoopTimer
over and over. Additionally, the CFRunLoopTimer isn't obviously doing any useful work.

This is expending power for a service worker that is mostly just idle.

We currently don't have a way to track down what CFRunLoopTimer is causing this, but we can detected
a service worker that has done nothing but timeout trying to handle a service worker task multiple times
in a row, and has done so for "defaultTerminationDelay" seconds.

This patch detects such misbehavior, logs it, causes a simulated fault with that info, and terminates the SW.

In a standalone followup I hope to change our RunLoop/RunLoopTimer implementation to track which timers are
installed and to be able to log useful information about them.

* Source/WebCore/workers/WorkerRunLoop.cpp:
(WebCore::WorkerDedicatedRunLoop::run): If the spin of the runloop results in multiple MessageQueueTimeout
  results in a row, and has done so for more than "defaultTerminationDelay" seconds, then terminate the SW.