Remove MallocPtr. https://bugs.webkit.org/show_bug.cgi?id=294022 rdar://152575154
Reviewed by NOBODY (OOPS!).

[This is a second submission of effectively the same patch. The first one was reverted
because new non-conforming uses of MallocPtr were added to the codebase while the original
PR was in review.]

This patch changes `MallocPtr` to require the parameter type `T` to be trivially destructible.

The usage of `MallocPtr` in `WasmTable` which was responsible for a callee leak is now
incompatible with that requirement. `WasmTable` is changed to use `std::unique_ptr`
with a custom deleter that correctly destroys the table contents.

Additionally, `VM` is changed to use a `FixedVector` instead of a `MallocPtr` to allocate
the exception fuzzing buffer.