AX: Legacy YouTube embed elements create infinite loops in accessibility https://bugs.webkit.org/show_bug.cgi?id=296387 rdar://156331833
Reviewed by NOBODY (OOPS!).

Accessibility elements are (in most cases) based off of the renderer. For
YouTube embed elements, due to some special-casing, the Embed parent and
the shadow dom child share the same renderer (see
HTMLPlugInElement::createElementRenderer). This breaks accessibility
assumptions. Instead, if we create a separate renderer for the plug in when
applicable, we can fix the broken accessibility structure while also
creating any renderering regressions.

This patch adds two new tests. The accessibility test verififes that we
no longer hang when buidling the accessibility tree. The layout test verifies
that the renderers for the embed and its child are unique.

* LayoutTests/accessibility/youtube-embed-accessibility-hierarchy-expected.txt: Added.
* LayoutTests/accessibility/youtube-embed-accessibility-hierarchy.html: Added.
* LayoutTests/dom/html/youtube-embed-rendering-expected.txt: Added.
* LayoutTests/dom/html/youtube-embed-rendering.html: Added.
* Source/WebCore/html/HTMLPlugInElement.cpp:
(WebCore::HTMLPlugInElement::createElementRenderer):