Event Timing: dispatch to PerformanceObserver https://bugs.webkit.org/show_bug.cgi?id=297010
Reviewed by NOBODY (OOPS!).

Implements the dispatch of PerformanceEventTiming entries to
PerformanceObsevers.

The use of maplike on EventCounts was commented in favor of
a manual implementation to avoid test flakiness.

Some code was moved around to better manage dependencies, and this
required some header changes: PerformanceEventTimingCandidate is now
PerformanceEventTiming::Candidate, and this required PerformanceEntry.h
to no longer include Performance.h in order to avoid a circular
dependency.

* LayoutTests/imported/w3c/web-platform-tests/event-timing/idlharness.any-expected.txt: now fails - missing maplike interface
* LayoutTests/imported/w3c/web-platform-tests/event-timing/idlharness.window-expected.txt: now fails - missing maplike interface
* Source/WebCore/Headers.cmake: fix build (removal of Performance.h from PerformanceEntry.h)
* Source/WebCore/WebCore.xcodeproj/project.pbxproj: fix build (removal of Performance.h from PerformanceEntry.h)
* Source/WebCore/page/PerformanceMark.h: fix build (removal of Performance.h from PerformanceEntry.h)
* Source/WebCore/page/PerformanceNavigationTiming.cpp: fix build (removal of Performance.h from PerformanceEntry.h)