Web Inspector: Created Accessibility Tree View https://bugs.webkit.org/show_bug.cgi?id=151137 rdar://23500025
Reviewed by NOBODY (OOPS!).

Added a toggle button on the Elements Tab view that toggles between displaying the accessibility tree and
 the DOMtree. It also toggles the details side panel that is displayed along with each content view.
For the accessibility tree, the accessibility details side panel is rendered. The accessibility
tree displays accessibility properties of each node such as the role, label, and focused
properties. For accessibility issues such as missing labels on images and links, it
displays a warning symbol that highlights this accessibility issues.

It builds upon the existing DOM Tree structure/architecture in order to maintain similar functionalities.