[Grid][Baseline] Fix RowAxisPositioning for baseline alignment in grid layout. https://bugs.webkit.org/show_bug.cgi?id=296058 <rdar://155967278>
Reviewed by NOBODY (OOPS!).

This PR sets the default alignment for first and last baseline
to `GridAxisPosition::GridAxisStart` and `GridAxisPosition::GridAxisEnd`

as per:

https://www.w3.org/TR/css-align-3/#baseline-values

`The fallback alignment for first baseline is safe self-start (for self-alignment)
or safe start (for content-distribution).

The fallback alignment for last baseline is safe self-end (for self-alignment)
or safe end (for content-distribution).`

This PR also handles opposite flow grid items to resolve to the
correct physical direction. Orthogonal flow grid items are set to
follow the grid's writing mode.

* LayoutTests/fast/grid/grid-baseline-alignment-orthogonal-flow-expected.html: Added.
* LayoutTests/fast/grid/grid-baseline-alignment-orthogonal-flow.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-justify-baseline-002-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-justify-baseline-003-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-justify-baseline-004-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-justify-baseline-005-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/subgrid/subgrid-baseline-009-expected.txt:
* Source/WebCore/rendering/RenderGrid.cpp:
(WebCore::RenderGrid::rowAxisPositionForGridItem const):