Use of MESSAGE_CHECK for stream connection messages is invalid https://bugs.webkit.org/show_bug.cgi?id=296697 rdar://157109453
Reviewed by NOBODY (OOPS!).

MESSAGE_CHECK calls IPC::Connection::markCurrentlyDispatchedMessageAsInvalid, which ASSERTs if it is called while not currently handling an IPC message. Stream connections dispatch and handle messages slightly differently and customise the behaviour of markCurrentlyDispatchedMessageAsInvalid to account for this. This patch introduces a STREAM_MESSAGE_CHECK_WITH_MESSAGE_BASE macro which accepts the StreamServerConnection rather than the underlying IPC::Connection, so the right invalidation behaviour can be used. This patch also addresses accounting for message dispatch of out of stream messages and adopts STREAM_MESSAGE_CHECK_WITH_MESSAGE_BASE in a few interfaces that were previously using it incorrectly.

* Source/WebKit/GPUProcess/graphics/RemoteImageBuffer.cpp:
* Source/WebKit/GPUProcess/graphics/RemoteImageBufferSet.cpp:
* Source/WebKit/GPUProcess/graphics/RemoteRenderingBackend.cpp:
* Source/WebKit/Platform/IPC/Connection.h:
* Source/WebKit/Platform/IPC/StreamServerConnection.cpp:
(IPC::StreamServerConnection::processOutOfStreamMessage):