document.caretPositionFromPoint over a video element returns null https://bugs.webkit.org/show_bug.cgi?id=287796 rdar://145477560
Reviewed by NOBODY (OOPS!).

`caretPositionFromPoint` works by performing hit-testing to find a node, getting
it's renderer, and then using `RenderObject::positionForPoint` to obtain a
`VisiblePosition`.

Currently this logic can incorrectly return `null` when the hit-tested node is a
video element. `RenderObject::createVisiblePosition` (via `RenderReplaced::positionForPoint`)
creates a legacy `Position` with the video element, an offset of 0, and the
anchor type `PositionIsOffsetInAnchor`. This is different from other replaced
elements, which would have an anchor type `PositionIsBeforeAnchor` for the offset
0. This difference comes from the fact that `canContainRangeEndPoint` returns
true for video (consequently `editingIgnoresContent` returns false), while it
does not for other replaced elements.

Ultimately, the difference above results in a null `VisiblePosition` in the case
where there is no next/previous candidate. For other replaced elements, the
`positionBeforeOrAfterNodeIsCandidate` check would succeed for the downstream
candidate, and that position would be returned.

Fix by updating `HTMLMediaElement` to return false from `canContainRangeEndPoint`.

* LayoutTests/imported/w3c/web-platform-tests/css/cssom/caretPositionFromPoint-with-audio-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/caretPositionFromPoint-with-audio.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/caretPositionFromPoint-with-video-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/caretPositionFromPoint-with-video.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/cssom/support/white.webm: Added.
* Source/WebCore/html/HTMLMediaElement.h:

The content inside the opening and closing video tags is only shown as a fallback
in engines that don't support the element. Consequently, the range endpoint can
never be inside the video element in WebKit.

The same is true of audio tags. To cover both elements, the method is overridden
on `HTMLMediaElement`.